pending

+++ banking-system-prototype/src/main/java/com/banking/prototype/repositories/BankAccountRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BankAccountRepository extends JpaRepository<BankAccount, Integer> {
+public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
     List<BankAccount> findAllByInfo_UserId(int userId);
 }
